refactor(server): per-platform capability registry for gateway connections
chat-instance-manager.ts (1885 -> 1397) no longer hardcodes platform names:
routing-info extraction, file handlers, instruction providers, webhook
secret/config and command registration are optional capabilities on
per-platform descriptor modules under gateway/connections/platforms/
(successor to ADAPTER_FACTORIES). Slack-only response-strategy utilities
moved out of the shared platform-strategies module. Adding a platform is now
one descriptor entry instead of edits across the manager. Logic moved
verbatim; public surface and behavior unchanged.
